backend/app/schema/vote.py
# -*- coding:utf-8 -*-
# Copyright (C) 2023  E.SUN BANK.
# @Author: Hsin-Hsien Ho, Wan-Chu Lin
# @Date: 2023/01
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.

# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.

from pydantic import BaseModel, Field


class VoteCreate(BaseModel):
    account: str = Field(..., title="帳號", example="admin")
    tid: int = Field(..., title="隊伍編號", example=1)
    gid: str = Field(..., title="組別", example="歌唱組")
