backend/requirements.txt
bcrypt==4.0.1
fastapi==0.78.0
gunicorn[gevent]==20.1.0
pandas==1.3.5
passlib==1.7.4
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
requests==2.27.1
sqlalchemy==1.4.36
openpyxl==3.0.10
uvicorn==0.17.6
